+++ GlTron/src/com/glTron/OpenGLView.java
@@ -22,23 +22,37 @@
 
 package com.glTron;
 
+import javax.microedition.khronos.egl.EGLConfig;
+import javax.microedition.khronos.opengles.GL10;
+
 import android.content.Context;
 import android.opengl.GLSurfaceView;
 import android.os.Handler;
 import android.view.MotionEvent;
 
 import com.glTron.logging.Logger;
+import com.glTron.Game.GLTronGame;
 
-public class OpenGLView extends GLSurfaceView {
+public class OpenGLView extends GLSurfaceView implements GLSurfaceView.Renderer 
+{
 	
-	private OpenGLRenderer _renderer;
+	//private OpenGLRenderer _renderer;
+	
+	GLTronGame Game;
+	Context mContext;
+	private int frameCount = 0;
 	
 	public OpenGLView(Context context,int width, int height) {
 		super(context);
 		Logger.v(this, "View's constructor called");
+
 		Logger.v(this, "Setting up the renderer object");
-		_renderer = new OpenGLRenderer(context, width, height);
-		setRenderer(_renderer);
+		//_renderer = new OpenGLRenderer(context, width, height);
+		//_renderer = new GLSurfaceView.Renderer();
+		setRenderer(this);
+
+		Game = new GLTronGame();
+		mContext = context;
 
 		Logger.v(this, "All set for View");
 	}
@@ -46,13 +60,15 @@ public class OpenGLView extends GLSurfaceView {
 	public void onPause()
 	{
 		Logger.v(this, "View Paused");
-		_renderer.onPause();
+		//_renderer.onPause();
+		Game.pauseGame();
 	}
 	
 	public void onResume()
 	{
 		Logger.v(this, "View Resumed");
-		_renderer.onResume();
+		//_renderer.onResume();
+		Game.resumeGame();
 	}
 	
 	public boolean onTouchEvent(final MotionEvent event) 
@@ -62,9 +78,41 @@ public class OpenGLView extends GLSurfaceView {
 		if(event.getAction() == MotionEvent.ACTION_DOWN) 
 		{
 			Logger.v(this, "Passing the touchevent to the renderer object");
-			_renderer.onTouch(event.getX(),event.getY());
+			//_renderer.onTouch(event.getX(),event.getY());
+			Game.addTouchEvent(event.getX(),event.getY());
 		}
 		
 		return true;
+	}
+
+	@Override
+	public void onSurfaceCreated(GL10 gl, EGLConfig config) 
+	{
+		Logger.Debug(this, "Surface Created, Do perspective");
+
+		Game.drawSplash(mContext, gl);
+	}
+	
+	@Override
+	public void onSurfaceChanged(GL10 gl, int w, int h) 
+	{
+		Logger.Debug(this, "Surface changed, Update the game screen");
+		Game.updateScreenSize(w, h);
+	}
+	
+	@Override
+	public void onDrawFrame(GL10 gl) 
+	{
+		if(frameCount == 1)
+		{ 
+			Logger.v(this, "Drawing the first frame for the game");
+			Game.initialiseGame(); 
+		}
+		else if(frameCount > 1) 
+		{ 
+			Game.RunGame(); 
+		}
+
+		frameCount++;
 	}
 }
